# specs/001-rag-recommendation-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Aetheris RAG 系统 API
  description: |
    面向高校学习场景的学习资源检索与推荐 RAG 系统接口文档。

    **核心功能**：
    - 用户认证（注册、登录）
    - 学习资源管理（上传、列表、详情）
    - RAG 智能问答（带引用来源）
    - 个性化推荐（基于用户画像）
    - 用户行为记录（查询、点击、收藏）
    - 离线评测（Precision@K、Recall@K）

  version: 1.0.0
  contact:
    name: Aetheris RAG Team
    email: support@aetheris.dev
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: https://api.aetheris.dev/api
    description: 生产环境
...
paths:
...
  /resources:
    post:
      tags: [resources]
      summary: 上传学习资源
      description: |
        上传 PDF 或 Markdown 文档，系统自动提取文本、切片、向量化入库。

        **幂等性保证**：基于文件内容哈希（SHA-256）去重，重复上传返回已存在的资源。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadResourceRequest'
      responses:
        '201':
          description: 上传成功，资源已入库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '400':
          description: 请求错误（文件格式不支持、文件过大、重复上传）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [resources]
      summary: 获取资源列表
      description: 分页获取资源列表，支持按标签过滤
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码（从 0 开始）
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 每页大小
          required: false
          schema:
            type: integer
            default: 20
        - name: tags
          in: query
          description: 标签过滤（逗号分隔，如：机器学习,深度学习）
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceListResponse'
  /resources/{id}:
    get:
      tags: [resources]
      summary: 获取资源详情
      description: 获取资源详细信息，包含切片列表
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 资源ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDetailResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UploadResourceRequest:
      type: object
      required:
        - file
        - title
        - fileType
      properties:
        file:
          type: string
          format: binary
          description: 文档文件（PDF 或 Markdown）
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: 深度学习基础教程
        tags:
          type: string
          description: 标签（逗号分隔）
          example: 机器学习,深度学习,神经网络
        description:
          type: string
          example: 本书介绍深度学习的基础概念和常用算法
    ResourceResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 深度学习基础教程
        tags:
          type: string
          example: 机器学习,深度学习
        fileType:
          type: string
          enum: [PDF, MARKDOWN]
          example: PDF
        fileSize:
          type: integer
          description: 文件大小（字节）
          example: 5242880
        description:
          type: string
          example: 本书介绍深度学习的基础概念
        uploadTime:
          type: string
          format: date-time
          example: '2025-12-25T10:30:00Z'
        chunkCount:
          type: integer
          description: 切片数量
          example: 150
        vectorized:
          type: boolean
          description: 是否已向量化
          example: true
    ResourceListResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ResourceResponse'
        total:
          type: integer
          description: 总数量
          example: 50
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
    Chunk:
      type: object
      properties:
        id:
          type: integer
          example: 1
        chunkIndex:
          type: integer
          description: 切片序号
          example: 5
        locationInfo:
          type: string
          description: 定位信息
          example: PDF 第 12-14 页
        snippet:
          type: string
          description: 文本摘录（前 200 字符）
          example: 深度学习是机器学习的一个子领域，它基于人工神经网络...
    ResourceDetailResponse:
      type: object
      properties:
        resource:
          $ref: '#/components/schemas/ResourceResponse'
        chunks:
          type: array
          description: 切片列表
          items:
            $ref: '#/components/schemas/Chunk'
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2025-12-25T10:30:00Z'
        status:
          type: integer
          example: 400
        error:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: 邮箱已存在
        path:
          type: string
          example: /api/auth/register